+++ GHEtool/gui/translation_class.py
@@ -216,6 +216,7 @@ class Translations:  # pragma: no cover
         "option_len_peak_heating",
         "option_len_peak_cooling",
         "option_temperature_profile_hourly",
+        "option_n_threads",
         "languages",
     )
 
@@ -1344,7 +1345,7 @@ class Translations:  # pragma: no cover
         self.option_pipe_depth: List[str] = [
             "Burial depth [m]:",
             "Vergrabungstiefe [m]:",
-            "Begraven diepte [m]:",
+            "Deklaag [m]:",
             "Burial depth [m]:",
             "Burial depth [m]:",
             "Burial depth [m]:",
@@ -1721,4 +1722,13 @@ class Translations:  # pragma: no cover
             "Should hourly data be used for the temperature profile?, no , yes ",
             "Should hourly data be used for the temperature profile?, no , yes ",
             "Should hourly data be used for the temperature profile?, no , yes",
+        ]
+        self.option_n_threads: List[str] = [
+            "Number of parallel threads [-]: ",
+            "Anzahl an parallelen Prozessen [-]: ",
+            "Nummer van parallelle threads [-]: ",
+            "Number of parallel threads [-]: ",
+            "Number of parallel threads [-]: ",
+            "Number of parallel threads [-]: ",
+            "Number of parallel threads [-]:",
         ]